Add files via upload
This is a Jupyter notebook for trying out the text analytics capability of scikit-learn, in particular using the 21newsgroup data. It basically follows the instructions given in the tutorial documentation: https://scikit-learn.org/stable/tutorial/text_analytics/working_with_text_data.html
There was some care needed to make sure that the relevant data sets are downloaded into the appropriate directories so that the code runs and produces results as expected.