Delete 20news1.ipynb
This was a very preliminary version of ds-mvp1 which was run with the 20news group data set from the scikit-learn text classification tutorial.  Since then, it's been re-factored and customized to work with the actual data for the taketwo project and the updated version now resides in "taketwo-dsmvp-v1.ipynb", so this file is no longer necessary.